All the functions are implemented. There is a bug.
Each function has been implemented and tested somewhat. Some
more functionality may need to be added depending on bugs.

There is a bug somewhere in the code that is causing
instructions to fire before they are actually can due to some
dependencies. I believe this has to do with the scheduling
method.